requirements.txt
#langchain
langchain
openai

#para usar o Tool SERP - a API grátis é bem limitada e o custo talvez não compense.
google-search-results

#biblioteca para converter qualquer arquivo não estruturado em um treco mais estruturado
unstructured

#pode ser usado para gerar embeddins localmente; mas agora estamos usando a api do openai pra isso tb
tiktoken

#weaviate (armazenamento de embeddings)
weaviate-client

#partition pdf
pdf2image
pdfminer.six

#etc
python-dotenv
